pyproject.toml
[tool.poetry]
name = "pygistwrapper"
version = "0.1.0"
description = "Interact with gist through API"
authors = ["Abhinav KM"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
anyio = "3.6.2"
black = "22.12.0"
certifi = "2022.12.7"
charset-normalizer = "3.0.1"
click = "8.1.3"
colorama = "0.4.6"
fastapi = "0.89.1"
h11 = "0.14.0"
idna = "3.4"
jinja2 = "3.1.2"
markupsafe = "2.1.2"
mypy-extensions = "0.4.3"
pathspec = "0.11.0"
platformdirs = "2.6.2"
pydantic = "1.10.4"
python-dateutil = "2.8.2"
python-dotenv = "0.21.1"
python-multipart = "0.0.6"
requests = "2.28.2"
six = "1.16.0"
sniffio = "1.3.0"
starlette = "0.22.0"
tomli = "2.0.1"
typing-extensions = "4.4.0"
urllib3 = "1.26.14"
uvicorn = "0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
